Add a database configuration DSL
This PR adds the ability to define database configurations in an
application using a Ruby DSL. This would allow applications with much
more complex configurations use Ruby over YAML. One of the advantages is
we can avoid making the YAML parsing for database configuration objects
more complex.

The defined behavior will use the `database.yml` file if present. If
there is no `database.yml` and there is a `database.rb` Rails will use
the Ruby file.

The structure of the Ruby config is very similar to the YAML. It
supports default settings that can get merged into the configurations to
avoid duplicating the same lines. It uses a dummy kwarg to avoid parsing
the configuration hash before boot just like the dummy database yaml.

To create a configuration DSL, delete your `database.yml` and add a
`database.rb`.

A database.yml that looks like this:

```yaml
default: &default
  adapter: mysql2
  encoding: utf8mb4
  pool: <%= ENV.fetch("RAILS_MAX_THREADS") { 5 } %>
  username: root
  password:
  socket: /tmp/mysql.sock

development:
  primary:
    <<: *default
    database: recipes_development
  primary_replica:
    <<: *default
    database: recipes_development
    replica: true
```

Would be written in the Ruby DSL like this:

```ruby
ActiveRecord::DatabaseConfigurations.configure do
  default = build do
    adapter "mysql2"
    encoding "utf8mb4"
    pool ENV.fetch("RAILS_MAX_THREADS") { 5 }
    username "root"
    password ""
    socket "/tmp/mysql.sock"
  end

  env(:development) do
    config(:primary, default) do
      database "recipes_development"

      replica do
        username "readonly_user"
      end
    end
  end
end
```

The Ruby example from above will return 2 configurations; one named
`primary` and one named `primary_replica`. If we passed a name to
`replica` like `replica(:one)` it would create a replica named
`primary_replica_one`. One of the benefits of using the Ruby DSL over
the YAML is we can ignore the default configuration. In YAML since we
don't know what someone will name their default we must always include
it the configurations list when calling
`ActiveRecord::Base.configurations`. When using the Ruby DSL we know
which ones are the default settings so it can be ignored.

If `dummy: true` is passed to the builder we'll skip evaluating the
block and return a list of configurations with `env_name` and `name`
set while `configuration_hash` will be empty.

In the future Rails should provide a generator option to use the Ruby
DSL over YAML and a task that can convert a YAML file into the Ruby DSL.
For now, this PR adds support for the DSL and replacement of the YAML
file will be manual.

The aim of this PR is not to replace YAML but to give an option for
more complex database configurations.

The main benefits of using the Ruby DSL are:

1) Can use simple Ruby to create multiple databases, ex:

```ruby
add(:primary) do
  database "my_database"

  10.times do |int|
    replica("#{int}") do
      username "readonly_user"
    end
  end
end
```

Or setup shards:

```ruby
add(:primary) do
  database "my_database"
  username "primary_user"

  shard(:one) do
    database "my_database_shard_one"
    username "shard_user"

    replica(:one) do
      username "shard_readonly_user"
    end

    replica(:two) do
      username "shard_readonly_user"
    end
  end
end
```

Or write out conditionals:

```ruby
add(:primary) do
  database Rails.config.some_value ? "my_database" : "your_database"
end
```

2) The Ruby DSL leaves the `default` config definitions out of the
`configurations` list. This isn't possible with YAML because we don't
know what the name of the default will be, it can be anything.

3) Reduces repetition by supporting nested configurations. This will
allow us to tie replicas to their primaries and eventually simplify
`connects_to` to only be `connects_to :primary`. This will never be
possible with the 3-tier YAML configuration.

4) The flexibility of YAML + ERB makes it difficult to "guess" where we
are in the levels. Adding a 4th level to YAML configurations for replicas
or shards is untenable without a stricter solution. The Ruby DSL makes it
easier to build complex configurations programtically.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>