[guides] Update missing translation example
The Rails Guides contain an example of how to raise for a missing
translation when `I18n.t` is called from anywhere, including
non-ActionView contexts. The example builds a custom exception handler
that catches and raises any instance of `I18n::MissingTranslationData`
and passes on any other exception to be handled by `I18n` as normally
happens.

Unfortunately, while `I18n::MissingTranslationData` is what eventually
should be raised for a missing translation, the initial exception the
handler sees is actually `I18n::MissingTranslation`, and
`MissingTranslationData` is the product of calling `#to_exception` on
that exception object. As a result, the example in the guides ends up
not raising for missing translations, silently resulting in the return
of the missing translation key.

This change updates the guides to with a working example, catching
`I18n::MissingTranslation` which results in a raise of
`I18n::MissingTranslationData` with the missing key.