Revert "Merge pull request #38658 from eileencodes/refactor-for-each-databases-code"
This reverts commit 9a5bc246d75595ae320a72c63a1df44ce6f63717, reversing
changes made to 227ff44e36ff9753fc5577335a333514949883fc.

I realized afterwards that this was written this way for a reason and it
wasn't accidental. The databases loaded in `for_each` are not complete
database configs because these are generated from the database yaml
before the Rails env is loaded. So we actually do need to refetch the
correct database configuration objects after we've loaded the Rails env.