Remove unnecessary creation of new hash config
This line was creating a new hash config, but we should always have a
hash config to lookup at this point.

Previously we needed this code because the pool name was merged into the
hash. Now it's an accessor on the db_config, so we have no reason to
create a new hash config.

Also, change env_name to name since it's only an env in a single db
application or an application with no arguments for
establish_connection.