Optimise length of default index name for polymorphic references
This reduces the possibility of generating index names that are too long
for certain databases (e.g. Postgres which has a 63 character limit) by
naming the index based on the reference name rather than the type and id
column names (the default behaviour of t.index).

This still allows the name to be explicitly specified by passing options.