Allows input IDs overriding in HTML5 valid format
Some form inputs generate their ID based on
the value of the name passed as an argument
to the helper. Nevertheless, this input
is in HTML4 valid format.

This change makes possible for users to
override the generated ID with the
options parameter.

This ID is checked to be in a HTML5 valid
format.