Add X-Original-To to mail using Mailgun recipient
Certain mail providers strip out the BCC field in emails, therefore,
ActionMailbox can miss valid messages.

However, the Mailgun provider does pass the BCC recipient as a parameter
in the inbound payload separate to the `body-mime`.

This PR is adds preliminary tests and code to start the conversation
around modifying the `raw_email`. Note: Some of the old tests now fail
due to the fact that I'm using the Mail gem to add the `X-Original-To`
header.

Advice welcome.

Related to #38446

CC: @jeremy and @kaspth