Remove mention of generating CoffeeScript files in Preprocessing section
of the Asset Pipeline guide

This section previously declared that the Rails controller generator
generated an `app/assets/javascripts/projects.coffee` and an
`app/assets/stylesheets/projects.scss` file. Then, in
ab123a33d2460dcdc5c36001cef5316eadc75fb3, we removed the assertion that
`app/assets/javascripts/projects.coffee` would be generated (to reflect
the switch to webpacker for managing javascript in Rails 6). However,
the preceding sentence and the subsequent paragraph were still written
with the assumption that a CoffeeScript file would be generated.

Now we've reworded those parts to be consistent with the fact that Rails
no longer generates a CoffeeScript file.

[ci skip]