Use monotonic time for duration calculations
Move the use of Time.now to monotonic time to calculate duration.
Monotonic time is preferable because in some cases Time.now can be unstable[1].
It could be set manually by a system administrator or an NTP daemon program.

Since we are already using ConcurrentRuby library, use the monotonic time from ConcurrentRuby.

The session‘s last_write_timestamp defaults to nil. We can use this
nil to know that we can write (because first time the value is not being set).
After first write, we use monotonic time minus last write to see if it‘s
bigger than the set send_to_replica_delay.

https://github.com/ruby-concurrency/concurrent-ruby/blob/aa9295ec46a97af67624a007b126971fd60b4d92/lib/concurrent-ruby/concurrent/utility/monotonic_time.rb

[1]: https://blog.dnsimple.com/2018/03/elapsed-time-with-ruby-the-right-way/