Ensure attributes are always inherited
Prior to this commit, attributes were not inherited by a subclass if
they were added to the parent class after the subclass had added its own
attributes.  This was due to the way attribute definitions were stored
in an append-only `class_attribute`.  This commit changes the way
attribute definitions are stored to ensure they are always inherited.