Ensure attribute decorations are always inherited
Prior to this commit, attribute decorations were not inherited by a
subclass if they were added to the parent class after the subclass had
added its own attribute decorations.  This was due to the way attribute
decorations were stored in an append-only `class_attribute`.  This
commit changes the way attribute decorations are stored to ensure they
are always inherited.