Soft deprecate the `database` kwarg in 6.0
The database kwarg is deprecated in 6.1 and will be removed in 6.2. It
has caused a lot of confusion and is dangerous to use in requests. Even
though docs recommended against use in requests, the majority of bug
reports to Rails regarding `connected_to` are related to this feature.
Since it's not an adequate replacement for sharding support we're
removing it. If you need shard support please use Rails 6.1 and the
shard kwarg.