Fix `assert_called_with` an array as expected argument
Despite `assert_called_with` is a private api in Rails
I decided to suggest this change for consideration.
Feel free to close this if it does not make sense.

See test cases in the git-diff and this issue for context:
https://github.com/bogdanvlviv/minitest-mock_expectations/issues/4

```ruby
\# Before:
assert_called_with(@object, :<<, [ [ [2] ] ]) do
  @object << [2]
end

\# After:
assert_called_with(@object, :<<, [ [2] ]) do
  @object << [2]
end
```