Changed partial example to be more rails common
If you create an action, you have some instance variable which are
defined like in this example should be `@products`. If it is empty, then
Products is an empty array: `@products = []` or an empty list. The
example before worked only if `@products = nil` which is a rare case for
such code.

Also the description of `empty_list.html.erb` partials says, that the
list is empty. For me this is `@products = []` and not `@products =
nil`. That's why I've changed the example to this.