Improves ActiveRecord::Base.connection_handler.while_preventing_writes documentation
Not all write queries which are wrapped in a while_preventing_writes block are
blocked. For example in Postgres these are valid queries which are actually
executed:

    ActiveRecord::Base.connection_handler.while_preventing_writes do
      ActiveRecord::Base.connection.execute "EXPLAIN ANALYZE DELETE FROM users WHERE id = 26"
      ActiveRecord::Base.connection.execute "/*/**/SELECT*/ UPDATE users SET role = 'admin' WHERE id = 10"
      ActiveRecord::Base.connection.execute "SELECT * INTO users_new FROM users"
    end

The documentation should mention, that this method is not supposed to catch all
writes, but rather the usual writes as they would be created by ActiveRecord.