Ensure `serialize` never raises an error
`serialize` shouldn't be used for validation and serialization. In all
cases in Rails this was true except for values that went through
`ensure_in_range` which would raise a `RangeError` that was subsequently
rescued.

Before serializing a value it should be checked whether the value is
serializable otherwise serializing may have unspecified results.

* `ensure_in_range?` has been removed
* This change implements `serializable?` to check whether a value is valid
for serialization.
* Added a new `empty` method on `ActiveRecord::Result` to return a
result with an empty set.
* Returns nil instead of raising when the value is not serializable.

Co-authored-by: Aaron Patterson <tenderlove@ruby-lang.org>