Allow unique in addition to uniq in migration generator
Running the migration generator marking an attribute as unique was
previously done using the `uniq` option, e.g. `username:string:uniq`.
Wrongly using `unique` as the option skipped generating the index,
which could lead to uniqueness validations not working. This commit
allows the generator to be called with the `unique` option as well.