Add ActiveRecord::Core.[]
This commit implements the `[]` method on ActiveRecord::Core.  The point
of this feature is to give developers access to Arel attribute objects
so that they can construct more complex queries easily.  For example:

```
Developer.where(Developer[:salary].gt(9000))
```

We can construct a query with more complex conditions than "==" or "!=".
Using this method also correctly refers to columns in other tables, so
applying query constraints in joins will "Just Work".