Avoid specifying ruby version twice
On a new Rails project, the ruby version is specified twice: `.ruby-version` file and `Gemfile`. 
This is clearly a duplication and is not necessary since the Gemfile can read it from the `.ruby-version` file. When updating the ruby version you will now need to only update it in one place.
This change loads the version from the `.ruby-version` file.