Enables `cache_classes` for test environment by default
Enables `cache_classes` for test environment even when using Spring and
removes setting for `cache_template_loading` since that will be now
`true` implicitly.

- This setting was disabled in de46eff due to Spring executing the
reloaders when files changed and triggering an error condition in
Zeitwerk, but f8cbd3d removed the watchers and fixed the problem.

- When caching classes is disabled, modifying code while the suite is
running can have unintended side effects. For example, when using a
multi-db setup, `#connect_to` will make the suite fail when a class is
reloaded. The reason is that transactional fixtures rely on keeping
connections alive, while each class reload forces a reconnection and
produces an error when trying to reacquire closed connections.

- Removes `cache_template_loading` since that will be implicitly
`true` as per:
    https://github.com/rails/rails/blob/88ee52f9d9cf2068bb400205e5a0c1d2e40169d1/actionview/lib/action_view/railtie.rb#L84-L88).

See an example project reproducing the problem at:
https://github.com/jorgemanrubia/rails-error-when-editing-test-files