Fix merging IN clause to be AND as Boolean algebra
Merging relation in Active Record behaves to replace a condition if the
condition is an equality, because `col = A AND col = B` is always empty
in that case, it is quite inconvenient in Active Record.

But IN clause is not an equality, but a set obviously.
So `col IN (A, B, ...) AND col (A, C, ...)` is not always empty.

I think it is considered a bug, caused by 87b6856, returning operator as
`:==` accidentally affected to IN node since IN node was introduced as a
subclass of Equality node before (I suppose it was also accidentally).

This fix is just to make IN node to Binary node, not an Equality node.

As a result, Active Record will regard IN clause as a set, not an
equality.

```ruby
david_and_mary = Author.where(id: [david, mary])
david_and_bob  = Author.where(id: [david, bob])

# Before
david_and_mary.merge(david_and_bob) # => [david, bob]
david_and_bob.merge(david_and_mary) # => [david, mary]

# After
david_and_mary.merge(david_and_bob) # => [david]
david_and_bob.merge(david_and_mary) # => [david]
```

Fixes #39232.