Move and update enlist_fixture_connections test
This moves the previous test into the old test and reuses the connection
that that test establishes rather than requiring we muck with temporary
connection pool.

The change here is more correct than the previous code since we're
establishing new connections we should be checking the newly established
reading and writing connections are the same, not checking against the
existing ActiveRecord::Base.connection.

The test here also most closely emulates a real application using
multiple databases.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>