Add test to reproduce join + pluck bug
The PR #39264 fixed `pluck` so that when plucking on a join the string
columns correctly returned casted values rather than uncasted values.

This led me to dig into other patterns to ensure the changes were
correct. I discovered that 6.0 has been not-casting values for string
pluck, plucks with Arel.sql and symbol plucks.

While string and arel.sql plucks are fixed, symbol plucks still return
uncasted values when called on the joined table. Note the columns are
correctly casted for the Author table, but not for the books table.

I'm not entirely sure how to fix this since Rails quotes the column name
if it's a symbol so it has no type caster to cast on. This is a problem
for all non-string values like boolean, enum, etc.